extract details
