+++ src/components/Panel/AdditionalTextData.tsx
@@ -1,19 +1,24 @@
 import { smallMobileAndUnder } from "@/constants";
-import type { OracleQueryUI } from "@/types";
 import Chevron from "public/assets/icons/chevron.svg";
+import type { ReactNode } from "react";
 import { useState } from "react";
 import type { CSSProperties } from "styled-components";
 import styled from "styled-components";
 
+interface Props {
+  description: ReactNode;
+  queryText: string | undefined;
+  queryTextHex: string | undefined;
+}
 export function AdditionalTextData({
   description,
   queryText,
   queryTextHex,
-}: OracleQueryUI) {
+}: Props) {
   const [showBytes, setShowBytes] = useState(false);
-  const hasDescription = description !== "" && description !== "0x";
-  const hasQueryText = queryText !== "" && queryText !== "0x";
-  const hasQueryTextHex = queryTextHex !== "" && queryTextHex !== "0x";
+  const hasDescription = !!description && description !== "0x";
+  const hasQueryText = !!queryText && queryText !== "0x";
+  const hasQueryTextHex = !!queryTextHex && queryTextHex !== "0x";
   const chevronRotation = showBytes ? 0 : 180;
 
   function toggleShowBytes() {
