+++ src/components/Panel/Panel.tsx
@@ -1,7 +1,6 @@
 import { Button, ConnectButton, DecimalInput, PanelBase } from "@/components";
 import {
   blueGrey500,
-  blueGrey700,
   connectWallet,
   red500,
   settled,
@@ -17,20 +16,18 @@ import {
   usePanelContext,
   usePrimaryPanelAction,
 } from "@/hooks";
-import AncillaryData from "public/assets/icons/ancillary-data.svg";
 import Pencil from "public/assets/icons/pencil.svg";
 import Settled from "public/assets/icons/settled.svg";
-import Timestamp from "public/assets/icons/timestamp.svg";
 import Warning from "public/assets/icons/warning.svg";
 import type { CSSProperties, ReactNode } from "react";
-import { Fragment, useEffect, useState } from "react";
+import { useEffect, useState } from "react";
 import styled from "styled-components";
 import { useAccount, useNetwork } from "wagmi";
 import { ActionDetails } from "./ActionDetails";
-import { AdditionalTextData } from "./AdditionalTextData";
+import { Details } from "./Details";
 import { InfoIcons } from "./InfoIcons";
 import { Title } from "./Title";
-import { Link, MessageLink, Text } from "./style";
+import { MessageLink, SectionTitle, SectionTitleWrapper, Text } from "./style";
 
 const messageBackgroundColor = addOpacityToHsla(blueGrey500, 0.05);
 const errorBackgroundColor = addOpacityToHsla(red500, 0.05);
@@ -45,15 +42,7 @@ export function Panel() {
   const [message, setMessage] = useState<ReactNode>("");
   const [proposePriceInput, setProposePriceInput] = useState("");
   const [inputError, setInputError] = useState("");
-  const {
-    chainId,
-    oracleType,
-    valueText,
-    timeUNIX,
-    timeUTC,
-    moreInformation,
-    actionType,
-  } = content ?? {};
+  const { chainId, oracleType, valueText, actionType } = content ?? {};
 
   const primaryAction = usePrimaryPanelAction({
     query: content,
@@ -200,52 +189,12 @@ export function Panel() {
         )}
       </ActionsWrapper>
       {content && <InfoIcons {...content} />}
-      <DetailsWrapper>
-        <DetailWrapper>
-          <SectionTitleWrapper>
-            <TimestampIcon />
-            <SectionTitle>Timestamp</SectionTitle>
-          </SectionTitleWrapper>
-          <Time>
-            <TimeFormat>UTC</TimeFormat> {timeUTC}
-          </Time>
-          <Time>
-            <TimeFormat>UNIX</TimeFormat> {timeUNIX}
-          </Time>
-        </DetailWrapper>
-        <DetailWrapper>
-          <SectionTitleWrapper>
-            <AncillaryDataIcon />
-            <SectionTitle>Additional Text Data</SectionTitle>
-          </SectionTitleWrapper>
-          {content && <AdditionalTextData {...content} />}
-        </DetailWrapper>
-        <DetailWrapper>
-          <SectionTitleWrapper>
-            <AncillaryDataIcon />
-            <SectionTitle>More information</SectionTitle>
-          </SectionTitleWrapper>
-          {moreInformation?.map(({ title, href, text }) => (
-            <Fragment key={title}>
-              <SectionSubTitle>{title}</SectionSubTitle>
-              <Link href={href} target="_blank">
-                {text}
-              </Link>
-            </Fragment>
-          ))}
-        </DetailWrapper>
-      </DetailsWrapper>
+      {content && <Details {...content} />}
     </PanelBase>
   );
 }
 
 // wrappers
-
-const SectionTitleWrapper = styled.div`
-  display: flex;
-  gap: 12px;
-  margin-bottom: 16px;
-`;
 
 const ActionsWrapper = styled.div`
   background: var(--grey-400);
@@ -255,21 +204,6 @@ const ActionsWrapper = styled.div`
 `;
 
 const ActionButtonWrapper = styled.div``;
-
-const DetailsWrapper = styled.div`
-  padding-inline: var(--padding-inline);
-  padding-bottom: 64px;
-`;
-
-const DetailWrapper = styled.div`
-  padding-bottom: 20px;
-  &:not(:first-child) {
-    padding-top: 22px;
-  }
-  &:not(:last-child) {
-    border-bottom: 1px solid ${addOpacityToHsla(blueGrey700, 0.25)};
-  }
-`;
 
 const ValueWrapper = styled.div`
   width: min(100%, 512px);
@@ -308,24 +242,6 @@ const MessageWrapper = styled(ErrorWrapper)`
 
 // titles
 
-const SectionTitle = styled.h2`
-  font: var(--body-md);
-  font-weight: 700;
-
-  span {
-    font-weight: 400;
-  }
-`;
-
-const SectionSubTitle = styled.h3`
-  font: var(--body-sm);
-  font-weight: 600;
-
-  &:not(:first-child) {
-    margin-top: 16px;
-  }
-`;
-
 // text
 
 const ValueText = styled(Text)`
@@ -334,13 +250,6 @@ const ValueText = styled(Text)`
   @media ${smallMobileAndUnder} {
     font: var(--body-sm);
   }
-`;
-
-const Time = styled(Text)``;
-
-const TimeFormat = styled.span`
-  display: inline-block;
-  margin-right: 32px;
 `;
 
 const ErrorText = styled(Text)`
@@ -356,10 +265,6 @@ const MessageText = styled(Text)`
 // icons
 
 const PencilIcon = styled(Pencil)``;
-
-const TimestampIcon = styled(Timestamp)``;
-
-const AncillaryDataIcon = styled(AncillaryData)``;
 
 const WarningIcon = styled(Warning)``;
