+++ src/components/Panel/style.tsx
@@ -26,3 +26,18 @@ export const Link = styled(NextLink)`
 export const MessageLink = styled(Link)`
   word-break: normal;
 `;
+
+export const SectionTitleWrapper = styled.div`
+  display: flex;
+  gap: 12px;
+  margin-bottom: 16px;
+`;
+
+export const SectionTitle = styled.h2`
+  font: var(--body-md);
+  font-weight: 700;
+
+  span {
+    font-weight: 400;
+  }
+`;
